initialize router connection as soon as possible
When a gateway is started, or the connection has been reset, we want it
to be ready to talk to packet routers as soon as possible.

This will let it downlink to class-C devices without needing to first
see an uplink.

And it allows us to know it's availability without again needing devices
around it transmitting.